alike ud_feedback. And we dont need to user justify, we regard the top link as valid and then remove it then it will generate new top link. But I think the cost of false positive may be have a great deal of negative effect.
this branch is used to deal with issue:[iTrace-2] dont use ud_feedback, regard link which has top IR value valid
the result is negative because the IR may cheat with us, and if we regard a negative link as positive it will has great side effect.
At before, we has tried see the highest score as true, and it is not positive.
Here, the end condition is not easy to set. Initially, we think when the bonus is very small, iterate end. However, in practice, the *bonus* will not suddenly drop, for we use the global max score and
it will not change. So the end condition is: we only believe the top *count* links, the *count* is set by us (0<=count<=allLinksAmount). We find that the more *count*, the more negative result.